Restructure fuzzer
